moved generator components to scenario
