+++ flow/scenarios/figure_eight.py
