+++ flow/scenarios/loop_merge.py
@@ -28,7 +28,6 @@ class TwoLoopsOneMergingScenario(Scenario):
 
     def __init__(self,
                  name,
-                 generator_class,
                  vehicles,
                  net_params,
                  initial_config=InitialConfig(),
@@ -70,8 +69,139 @@ class TwoLoopsOneMergingScenario(Scenario):
         length_loop = 2 * pi * radius
         self.length_loop = length_loop
 
-        super().__init__(name, generator_class, vehicles, net_params,
-                         initial_config, traffic_lights)
+        super().__init__(name, vehicles, net_params, initial_config,
+                         traffic_lights)
+
+    def specify_nodes(self, net_params):
+        """See parent class."""
+        r = net_params.additional_params["ring_radius"]
+        x = net_params.additional_params["lane_length"]
+
+        nodes = [{
+            "id": "top_left",
+            "x": repr(0),
+            "y": repr(r),
+            "type": "priority"
+        }, {
+            "id": "bottom_left",
+            "x": repr(0),
+            "y": repr(-r),
+            "type": "priority"
+        }, {
+            "id": "top_right",
+            "x": repr(x),
+            "y": repr(r),
+            "type": "priority"
+        }, {
+            "id": "bottom_right",
+            "x": repr(x),
+            "y": repr(-r),
+            "type": "priority"
+        }]
+
+        return nodes
+
+    def specify_edges(self, net_params):
+        """See parent class."""
+        r = net_params.additional_params["ring_radius"]
+        x = net_params.additional_params["lane_length"]
+
+        ring_edgelen = pi * r
+        resolution = 40
+
+        edges = [{
+            "id":
+            "center",
+            "from":
+            "bottom_left",
+            "to":
+            "top_left",
+            "type":
+            "edgeType",
+            "length":
+            repr(ring_edgelen),
+            "priority":
+            "46",
+            "shape":
+            " ".join([
+                "%.2f,%.2f" % (r * cos(t), r * sin(t))
+                for t in linspace(-pi / 2, pi / 2, resolution)
+            ]),
+            "numLanes":
+            str(self.inner_lanes)
+        }, {
+            "id": "top",
+            "from": "top_right",
+            "to": "top_left",
+            "type": "edgeType",
+            "length": repr(x),
+            "priority": "46",
+            "numLanes": str(self.outer_lanes)
+        }, {
+            "id": "bottom",
+            "from": "bottom_left",
+            "to": "bottom_right",
+            "type": "edgeType",
+            "length": repr(x),
+            "numLanes": str(self.outer_lanes)
+        }, {
+            "id":
+            "left",
+            "from":
+            "top_left",
+            "to":
+            "bottom_left",
+            "type":
+            "edgeType",
+            "length":
+            repr(ring_edgelen),
+            "shape":
+            " ".join([
+                "%.2f,%.2f" % (r * cos(t), r * sin(t))
+                for t in linspace(pi / 2, 3 * pi / 2, resolution)
+            ]),
+            "numLanes":
+            str(self.inner_lanes)
+        }, {
+            "id":
+            "right",
+            "from":
+            "bottom_right",
+            "to":
+            "top_right",
+            "type":
+            "edgeType",
+            "length":
+            repr(ring_edgelen),
+            "shape":
+            " ".join([
+                "%.2f,%.2f" % (x + r * cos(t), r * sin(t))
+                for t in linspace(-pi / 2, pi / 2, resolution)
+            ]),
+            "numLanes":
+            str(self.outer_lanes)
+        }]
+
+        return edges
+
+    def specify_types(self, net_params):
+        """See parent class."""
+        speed_limit = net_params.additional_params["speed_limit"]
+
+        types = [{"id": "edgeType", "speed": repr(speed_limit)}]
+        return types
+
+    def specify_routes(self, net_params):
+        """See parent class."""
+        rts = {
+            "top": ["top", "left", "bottom", "right", "top"],
+            "bottom": ["bottom", "right", "top", "left", "bottom"],
+            "right": ["right", "top", "left", "bottom"],
+            "left": ["left", "center", "left"],
+            "center": ["center", "left", "center"]
+        }
+
+        return rts
 
     def specify_edge_starts(self):
         """See parent class."""
